remove default version of GmsCompatConfig dependency
Dependencies are always updated to their latest version before installation or update of their
dependant.
The only case when GmsCompatConfig version needs to be manually specified is if it's higher
than what is in stable channel, to make sure it's updated to alpha or beta version regardless
of the release channel selection.